remove redux-devtools
