+++ client/src/store/configureStore.dev.js
@@ -3,16 +3,12 @@ import thunk from 'redux-thunk'
 import createLogger from 'redux-logger'
 import api from '../middleware/api'
 import rootReducer from '../reducers'
-import DevTools from '../containers/DevTools'
 
 export default function configureStore(initialState) {
 	const store = createStore(
 		rootReducer,
 		initialState,
-		compose(
-			applyMiddleware(thunk, api, createLogger()),
-			DevTools.instrument()
-		)
+		applyMiddleware(thunk, api, createLogger())
 	);
 
 	if (module.hot) {
